Implemented a `Railway` controller.
Main purpose of this class is to execute, to interpret and to handle all
the informations about `Path`s, `PathInfo`s and `Section`s data.
(Need a renaming). By now, it with the function `getTripSolutions()`
it's possible to specify two stations names (String) and a result of
type `TripSolution` is yield. It's necessary to create on top of this
result a javafx `TableView` for displaying the results on the GUI.