Changed attributes on `PERCORRENZA`.
Added `tempoEffettivo` and `postiDisponibili`, and thanks to them it is
possibile to model a "dynamic" database. If the train is RV, the
`tempoEffettivo` field will be decremented by a certain amount (because
RV service offers higher speed trains); moreover, a passenger that buys
a ticket, affects the paramenter `postiDisponibili` by decrementing them
by one.